refacts

+++ src/engine/util/coord.rs
@@ -2,22 +2,22 @@ use crate::Size2D;
 
 #[derive(Copy, Clone, Debug)]
 pub struct ScreenCoord {
-   pub x: i32,
-   pub y: i32,
+   pub x: f64,
+   pub y: f64,
 }
 
 impl ScreenCoord {
    pub fn empty() -> ScreenCoord {
-      ScreenCoord { x: 0, y: 0 }
+      ScreenCoord { x: 0.0, y: 0.0 }
    }
-   pub fn from(x: i32, y: i32) -> Self {
+   pub fn from(x: f64, y: f64) -> Self {
       Self { x, y }
    }
-   pub fn from_tup((x, y): (i32, i32)) -> Self {
+   pub fn from_tup((x, y): (f64, f64)) -> Self {
       Self { x, y }
    }
    pub fn is_inside(&self, size: Size2D) -> bool {
-      if (self.x >= 0 && self.y >= 0) && (self.x <= size.w as i32 && self.y <= size.h as i32) {
+      if (self.x >= 0.0 && self.y >= 0.0) && (self.x <= size.w as f64 && self.y <= size.h as f64) {
          true
       } else {
          false
@@ -27,18 +27,18 @@ impl ScreenCoord {
 
 #[derive(Copy, Clone, Debug)]
 pub struct ScreenOffset {
-   pub x: i32,
-   pub y: i32,
+   pub x: f64,
+   pub y: f64,
 }
 
 impl ScreenOffset {
    pub fn empty() -> ScreenOffset {
-      ScreenOffset { x: 0, y: 0 }
+      ScreenOffset { x: 0.0, y: 0.0 }
    }
-   pub fn from(x: i32, y: i32) -> Self {
+   pub fn from(x: f64, y: f64) -> Self {
       Self { x, y }
    }
-   pub fn from_tup((x, y): (i32, i32)) -> Self {
+   pub fn from_tup((x, y): (f64, f64)) -> Self {
       Self { x, y }
    }
 }
